refactor: move cpuprofile test helpers to separate module
